mv ReadChainConfig & WriteChainConfig to core
