+++ cmd/rpcdaemon/cli/config.go
@@ -65,6 +65,7 @@ import (
 	"github.com/erigontech/erigon/cmd/rpcdaemon/rpcservices"
 	"github.com/erigontech/erigon/cmd/utils"
 	"github.com/erigontech/erigon/cmd/utils/flags"
+	"github.com/erigontech/erigon/core"
 	"github.com/erigontech/erigon/core/state"
 	"github.com/erigontech/erigon/core/tracing"
 	"github.com/erigontech/erigon/core/vm/evmtypes"
@@ -397,7 +398,7 @@ func RemoteServices(ctx context.Context, cfg *httpcfg.HttpCfg, logger log.Logger
 			if err != nil {
 				return err
 			}
-			cc, err = rawdb.ReadChainConfig(tx, genesisHash)
+			cc, err = core.ReadChainConfig(tx, genesisHash)
 			if err != nil {
 				return err
 			}
@@ -1016,7 +1017,7 @@ func (e *remoteConsensusEngine) init(db kv.RoDB, blockReader services.FullBlockR
 		if err != nil {
 			return err
 		}
-		cc, err = rawdb.ReadChainConfig(tx, genesisHash)
+		cc, err = core.ReadChainConfig(tx, genesisHash)
 		if err != nil {
 			return err
 		}
